refactor: create some new component
